requirements.txt
# Discord Bot - Versão mais recente com suporte a novos recursos
discord.py>=2.5.0

# HTTP Client
httpx>=0.27.0
aiohttp>=3.9.0

# OpenAI API
openai>=1.30.0

# Database
aiosqlite>=0.20.0

# Configuration
pyyaml>=6.0.1
python-dotenv>=1.0.0

# Vector Stores (opcional)
pinecone-client>=3.0.0
qdrant-client>=1.8.0
redis>=5.0.0
pgvector>=0.2.5
psycopg2-binary>=2.9.9

# LangChain (opcional)
langchain>=0.2.0
langchain-community>=0.2.0
langchain-openai>=0.1.0

# PDF Processing
pypdf>=4.0.0

# Utilities
numpy>=1.26.0
rapidfuzz>=3.6.0

# Web Server (opcional)
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
